fastapi-agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
